+++ error-handler/get-celebrate-error.js
@@ -1,0 +1,7 @@
+const responseMessages = require('../libs/response-messages');
+
+module.exports = (err, res) => {
+  return res
+    .status(400)
+    .send({ message: `${responseMessages.clientErrors.validation}: ${err.message}` });
+};
