+++ error-handler/index.js
@@ -1,14 +1,12 @@
 /* eslint no-unused-vars: 0   */
 const { isCelebrate } = require('celebrate');
 
+const getCelebrateErr = require('./get-celebrate-error');
 const translateErrors = require('./translate-errors');
-const responseMessages = require('../libs/response-messages');
 
 module.exports = (err, _req, res, next) => {
   if (isCelebrate(err)) {
-    return res
-      .status(400)
-      .send({ message: `${responseMessages.clientErrors.validation}: ${err.message}` });
+    return getCelebrateErr(err, res);
   }
   if (!err.statusCode) {
     return translateErrors(err, res);
