+++ errors/index.js
@@ -1,0 +1,9 @@
+const RequestWrong = require('./request-wrong');
+const ForbiddenError = require('./forbidden-error');
+const NotFoundError = require('./not-found-error');
+
+module.exports = {
+  RequestWrong,
+  ForbiddenError,
+  NotFoundError
+};
