+++ middlewares/not-found-res.js
@@ -1,5 +1,5 @@
 const responseMessages = require('../libs/response-messages');
-const NotFoundError = require('../errors/not-found-error');
+const { NotFoundError } = require('../errors');
 
 module.exports.notFoundRes = () => {
   throw new NotFoundError(responseMessages.clientErrors.resourceNotFound);
